Delete a particular resume

+++ serverMiddleware/users/deleteUser.js
@@ -12,14 +12,16 @@ app.use(bodyParser.json());
 module.exports = app.use(async function (req, res) {
     try {
 		const userToDelete = req.body.userId;
-		const visitorsToDelete = req.body.visitorsToDelete;
+        const visitorsToDelete = req.body.visitorsToDelete;
 
 		// 1) Delete all visitors accounts
 		for (let visitor of visitorsToDelete) {
 			await admin.auth().deleteUser(visitor)
 		}
-		// 2) Delete user account
-		await admin.auth().deleteUser(userToDelete)
+        // 2) Delete user account
+        if (userToDelete) {
+            await admin.auth().deleteUser(userToDelete)
+        }
         
         res.send({
             message: 'POST request to delete user account went successfully.'
